Screeplot

Creates a screeplot that that provides a quality measurement of the quality of the clustering of each SOM cell for both PAM and hierarchical clustering.

aweSOMscreeplot(
  ok.som,
  nclass = 2,
  method = "hierarchical",
  hmethod = "ward.D2"
)

Arguments

ok.som
SOM data object

nclass
number of superclasses to be visualized in the screeplot. Default is 2.

method
Method used for clustering. Hierarchical clustering ("hierarchical") and PAM ("pam") clustering can be used. By default hierarchical clustering is applied.

hmethod
Specifically for hierarchicical clustering, the clustering method can be specified. By default "ward.D2" is used for other options, see the stats::hclust method documentation.

Value

Screeplot

Examples

#' ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]

#> Error in eval(expr, envir, enclos): object 'ok.data' not found

### Scale training data
dat <- scale(dat)

#> Error in scale(dat): object 'dat' not found

## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)

#> Error in prcomp(traindat, center = F, scale. = F): object 'dat' not found

ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')

#> Error in check.data(data): object 'dat' not found

## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)

#> Error in inherits(x, "dist"): object 'ok.som' not found

superclasses <- unname(superclust$clustering)

#> Error in unname(superclust$clustering): object 'superclust' not found

aweSOM::aweSOMscreeplot(ok.som, method = 'hierarchical', hmethod = 'ward.D2', nclass = 2)

#> Error in aweSOM::aweSOMscreeplot(ok.som, method = "hierarchical", hmethod = "ward.D2",     nclass = 2): object 'ok.som' not found