D3-based plot data generation Function generates a list object containing data for D3-based plots

D3-based plot data generation Function generates a list object containing data for D3-based plots

getPlotParams(
  type,
  som,
  superclass,
  data,
  plotsize,
  varnames,
  normtype = c("range", "contrast"),
  palsc,
  palplot,
  cellNames,
  plotOutliers,
  reversePal,
  options = NULL,
  valueFormat
)

Arguments

type

som

superclass

data

plotsize

varnames

normtype

palsc

palplot

cellNames

plotOutliers

reversePal

options

valueFormat

Value