Import .xlsx data based on read_excel readxl

Import .xlsx data based on read_excel readxl

ok.data.function.excel_xlsx(
  input_dataFile,
  input_column_names,
  input_trim_spaces,
  input_range_specified_bol,
  input_range_specs,
  input_worksheet_specified_bol,
  input_worksheet_specs,
  input_dataFile_datapath,
  input_rows_to_skip
)

Arguments

input_dataFile
shiny-related dataFile object

input_column_names
TRUE to use the first row as column names, FALSE to get default names

input_trim_spaces
Should leading and trailing whitespace be trimmed?

input_range_specified_bol
boolean to indicate if range should be specified

input_range_specs
A cell range to read from. See documentation ?read_xlsx

input_worksheet_specified_bol
boolean to indicate if worksheet should be specified

input_worksheet_specs
Sheet to read. See documentation ?read_xlsx

input_dataFile_datapath
datapath

input_rows_to_skip
Minimum number of rows to skip before reading anything. Default is 0

Value